\name{FitVariogram}

\alias{FitVariogram}

\title{Fit Variogram}

\description{
Something.
}

\usage{
FitVariogram(obs, model, ...)
}

\arguments{
\item{obs}{SpatialPointsDataFrame; something.}
\item{model}{variogramModel; something.}
\item{...}{Arguments to be passed to \code{\link{fit.variogram}}.}
}

\details{
Something...
}

\value{
Something
}

\author{J.C. Fisher}

\seealso{
\code{\link{fit.variogram}}
}

\examples{
\donttest{
print("something")
}
}

\keyword{misc}
