\name{FitVariogram}

\alias{FitVariogram}

\title{Fit Variogram}

\description{
Something.
}

\usage{
FitVariogram(obs, model, ...)
}

\arguments{
\item{obs}{something.}
\item{model}{something}
\item{...}{something}
}

\details{
Something...
}

\value{
Something
}

\author{J.C. Fisher}

\seealso{
\code{\link{plot}}
}

\examples{
\donttest{
print("something")
}
}

\keyword{misc}
