\name{PlotKriging}

\alias{PlotKriging}

\title{Plot Kriging}

\description{
Something.
}

\usage{
PlotKriging(obs, v.fit, grd, rm.idxs, at.pred, at.se, debug.level = 0)
}

\arguments{
\item{obs}{something}
\item{v.fit}{something}
\item{grd}{something}
\item{rm.idxs}{something}
\item{at.pred}{something}
\item{at.se}{something}
\item{debug.level}{something}
}

\details{
Something...
}

\value{
Something
}

\author{J.C. Fisher}

\seealso{
\code{\link{plot}}
}

\examples{
\donttest{
print("something")
}
}

\keyword{misc}
