\name{PlotRaster}

\alias{PlotRaster}

\title{Create a Filled Contour Plot of Spatial Data}

\description{
Something.
}

\usage{
PlotRaster(grd, zcol, pts, ply, rm.idxs, xlim, ylim, at,
           pal = terrain.colors, contour = TRUE, label.pts = FALSE,
           main = "", gr.type = "windows", gr.file = NULL)
}

\arguments{
\item{grd}{something}
\item{zcol}{something}
\item{pts}{something}
\item{ply}{something}
\item{rm.idxs}{something}
\item{xlim, ylim}{something}
\item{at}{something}
\item{pal}{something}
\item{contour}{something}
\item{label.pts}{something}
\item{main}{something}
\item{gr.type}{something}
\item{gr.file}{something}
}

\details{
Something...
}

\value{
Something
}

\author{J.C. Fisher}

\seealso{
\code{\link{plot}}
}

\examples{
\donttest{
print("something")
}
}

\keyword{misc}
