\name{ProcessObservations}

\alias{ProcessObservations}

\title{Process Observation Data}

\description{
Something.
}

\usage{
ProcessObservations(d, x.var, y.var, projargs, siteno.var, sitenm.var,
                    net.var, var1.var, var2.var, acy.var, dt.var, dt.lim,
                    dt.fmt = "\%Y-\%m-\%d \%H:\%M")
}

\arguments{
\item{d}{something}
\item{x.var, y.var}{something}
\item{projargs}{something}
\item{siteno.var}{something}
\item{sitenm.var}{something}
\item{net.var}{something}
\item{var1.var}{something}
\item{var2.var}{something}
\item{acy.var}{something}
\item{dt.var}{something}
\item{dt.lim}{something}
\item{dt.fmt}{something}
}

\details{
Something...
}

\value{
Something
}

\author{J.C. Fisher}

\seealso{
\code{\link{plot}}
}

\examples{
\donttest{
print("something")
}
}

\keyword{misc}
