\name{ReadObservations}

\alias{ReadObservations}

\title{Read Observations}

\description{
Something.
}

\usage{
ReadObservations(file, x.var, y.var, site.var, obs.var, acc.var,
                 dt.var, dt.fmt = "\%Y\%m\%d",
                 projargs = "+proj=longlat +datum=NAD83")
}

\arguments{
\item{file}{something}
\item{x.var, y.var}{something}
\item{site.var}{something}
\item{obs.var}{something}
\item{acc.var}{something}
\item{dt.var}{something}
\item{dt.fmt}{something}
\item{projargs}{something}
}

\details{
Something...
}

\value{
Something
}

\author{J.C. Fisher}

\seealso{
\code{\link{plot}}
}

\examples{
\donttest{
print("something")
}
}

\keyword{misc}
