\name{RunCrossValidation}

\alias{RunCrossValidation}

\title{Run Crossvalidation}

\description{
Something.
}

\usage{
RunCrossValidation(formula, obs, grd, vg.model, nmax = Inf, ply = NULL)
}

\arguments{
\item{formula}{something}
\item{obs}{something}
\item{grd}{something}
\item{vg.model}{something}
\item{nmax}{something}
\item{ply}{something}
}

\details{
Something...
}

\value{
Something
}

\author{J.C. Fisher}

\seealso{
\code{\link{plot}}
}

\examples{
\donttest{
print("something")
}
}

\keyword{misc}
