\name{RunCrossvalidation}

\alias{RunCrossvalidation}

\title{Run Crossvalidation}

\description{
Something.
}

\usage{
RunCrossvalidation(obs, v.fit)
}

\arguments{
\item{obs}{something}
\item{v.fit}{something}
}

\details{
Something...
}

\value{
Something
}

\author{J.C. Fisher}

\seealso{
\code{\link{plot}}
}

\examples{
\donttest{
print("something")
}
}

\keyword{misc}
