\name{WriteGAResults}

\alias{WriteGAResults}

\title{Write Results of Genetic Algorithm}

\description{
Something.
}

\usage{
WriteGAResults(ga, file)
}

\arguments{
\item{ga}{something}
\item{file}{something}
}

\details{
Something...
}

\value{
Something
}

\author{J.C. Fisher}

\seealso{
\code{\link{plot}}
}

\examples{
\donttest{
print("something")
}
}

\keyword{IO}
